Fix render page Customized from API
This PR fixes #90. I have to admit, this page looks under in development
yet, and not ready to show the purpose of the `Customized` component.
However, I am still thinking is better to show the page than just a
broken page.

Notice the description is the same as Cell, and it's not clear the
function of this component. If you wish I can change it just to 'work in
progress' in the description.